requirements.txt
blinker==1.4
flask-bootstrap==3.3.7.1
flask-nav==0.6
flask-debug==0.4.3
flask-wtf==0.14.2
sqlalchemy==1.1.14
flask_sqlalchemy==2.2
Flask-Script==2.0.5
flask_migrate==2.1.1
Flask-Admin==1.5.0
Flask-Security==1.7.5
SQLAlchemy-Utils==0.32.16
raven==6.1.0
XlsxWriter==1.0.2
psycopg2==2.7.4
